Move model content to new access.py module
